split GUiLD & Sample Project into 2 components.
The COPYRIGHT.txt file now treats GUiLD and the included sample project
as two separate components, to make it more clear that the sample
project is optional and not a part of GUiLD itself.

This commit also renames the project to Godot Unofficial License Dialog,
because the old name was stupid.