remove _label_text variable.
The more I think about it, the more having a whole separate variable and
functions which modify the text of a label seemed wasteful and
unnecessary.  This commit removes that variable and repurposes the
`set_label_text` function as a shortcut.

Because this variable is no longer being used to modify the text of the
label, the script for the `LicenseDialog` no longer needs to update
editor properties, so the `tool` keyword has been removed.