do not read copyright file if path is blank.
If the `copyright_file` variable is left blank, the `LicenseDialog` will
not attempt to read any copyright file and will not populate the
`ComponentList` tree with project component information.

Projects will no longer push warnings if the `copyright_file` variable
is left as a blank string, which is now default.  However, the warnings
will still occur if the user specifies a file path which does not exist,
or if the project finds a copyright file which is not formatted
properly.

In order to make this change work, the `onready` keywords have been
removed from the `project_name` and `copyright_file` variable
declarations.